# perf: Use logical expression instead of regexp for `isUnitTestSetUp`