## Description

The `isUnitTestSetUp` function uses the regular expression `/^(?:before|after)(?:Each|All)$/`. This regex matches the following four strings:

- `beforeEach`
- `beforeAll`
- `afterEach`
- `afterAll`

When the options are limited to about four, using a `Set` is generally faster than performing calculations with a regular expression.

Below are the total execution times of the `isUnitTestSetUp` function against the source code of Babel and Prettier, comparing the `main` branch to this PR:


|repo |  `main` | this PR |
|--------|-------|---|
|prettier | 1.0767030000051818 ms   | 0.8237229999981537 ms |
|babel | 2.696221000274363 ms | 1.9283900000140193 ms |



## Checklist

<!-- Please ensure you’ve done all of these things (if applicable). -->
<!-- You can replace the `[ ]` with `[x]` to mark each task as done. -->

- [x] I’ve read the [contributing guidelines](https://github.com/prettier/prettier/blob/main/CONTRIBUTING.md).

<!-- Please DO NOT remove the playground link -->

**✨[Try the playground for this PR](https://prettier.io/playground-redirect)✨**