@bakkot 2024-02-05
If this is perf-critical, you should consider straight comparisons instead: `node.callee.name === 'beforeEach' || node.callee.name === 'beforeAll' || node.callee.name === 'afterEach' || node.callee.name === 'afterAll'`.

That's usually going to be faster than set lookup until you have quite a few items you're comparing against.