@fisker 2024-02-08
Due to the difference between babel and other parsers, the `isTestCall` result differently for optional call, but out scope of this PR.

**Prettier pr-16035**
[Playground link](https://deploy-preview-16035--prettier.netlify.app/playground/#N4Igxg9gdgLgprEAuEATOBnMAnAlgIzgH4A6ACgB0QMIBbOAAgxjygHMm7HnWOb6mLXO04Cewvl0G9R3ISP5yZi6RKoAaBmVTQ4ASgYBeAHwNgAXz0h1ICAAcYuaBmSgAhtmwQA7gAUPCC4obgA23m4Ani42+NhuYADWcDAAym70ADLCcMgAZqEYcDFxickpdvESyCwArkXUcLS41dh1NnAAHnZwePSwoQAqPVAeuJh5BfUYEiFwAIo1EPATIYU2AFYYHSkz84vLSPmr9QCO+3C+XnZBIG4YALRQcHDoqNYgLG64IRIAwnS0NzIW4hELvabsWYAQRgQnwNXgvh6WSeKzWIAAFjBaCEAOoY3DwDAVMBwFKBQm4ABuhIiwLAGGiICpdQAklB0LAUjhcA4oRyUjAIrM0fU7F5Cri4nZgeLMD0qTkbMJCtgYJc3GxAaKbBVsKrgULulg8A53uLhDBcbhUDAMcgABwABhs2DgZ1wbo1WqBh0mNhgbnw1tt9qQACYbDVCgMg0EjujGoRUK8Mm52DVNXAAGIQbCA2FVYIIiAgczmIA)
<!-- prettier-ignore -->
```sh
--parser typescript
```

**Input:**
<!-- prettier-ignore -->
```tsx
describe?.("some string some string some string some string some string some string some string some string", (done) => {})
```

**Output:**
<!-- prettier-ignore -->
```tsx
describe?.("some string some string some string some string some string some string some string some string", (done) => {});

```